Do not uncomment ccd directory.
Openvpn server can read and obey client specific configuration
directives in a directory, normally called 'ccd'. This is explained
in the configuration file, which we use. We do not create or use
client specific configurations.

In the upstream sample config file, this example is commented out:

https://github.com/OpenVPN/openvpn/blob/master/sample/sample-config-files/server.conf#L167

For some reason, this hasn't been commented out in our config.
This doesn't cause any problems on our Precise test and prod servers
running precise, but on the Xenial version the server doesn't start
and produces an error::

    Options error: --client-config-dir fails with 'ccd': No such file or directory
    Options error: Please correct these errors.

Change-Id: If00dfa2444bd540bcf3b28c20086dd2844d21a4b